# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Web3bridge Impact API",
    "description": "API for Web3bridge impact project v1",
    "contact": {
      "name": "API Support",
      "email": "mayowaobi74@gmail.com"
    }
  },
  "servers": [
    {
      "url": "https://impact-web3bridge-production.up.railway.app",
      "description": "This is the current deployed server url"
    }
  ],
...
  "securityDefinitions": {},
  "paths": {
...
    "/api/user/": {
      "get": {
        "tags": ["User"],
        "description": "You can add queries to the url to filter the response e.g. /api/user?cohortId=Cohort I&requestStatus=pending&role=super",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/{userId}": {
      "get": {
        "tags": ["User"],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["User"],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["User"],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}